requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.4.0
django-user-agents==0.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
lxml==4.5.2
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.1
